+++ django_athm/templatetags/athm_response_signal.py
@@ -11,13 +11,13 @@ logger = logging.getLogger(__name__)
 
 @register.simple_tag
 def athm_response_signal(status):
-    logger.debug("[templatetags:athm_response_signal]")
+    logger.debug("[django_athm:athm_response_signal]", extra={"status": status})
 
     if status == constants.EXPIRED_STATUS:
         signals.athm_expired_response.send(sender="django_athm")
     elif status == constants.CANCELLED_STATUS:
         signals.athm_cancelled_response.send(sender="django_athm")
-    elif status == constants.COMPLETED_STATUS:
+    elif status == constants.TRANSACTION_STATUS.COMPLETED:
         signals.athm_completed_response.send(sender="django_athm")
 
     # Send a general signal