+++ django_athm/utils.py
@@ -1,4 +1,5 @@
 import logging
+from abc import ABC, abstractmethod
 
 import httpx
 
@@ -11,30 +12,30 @@ def parse_error_code(error_code):
     return ERROR_DICT.get(error_code, "unknown error")
 
 
-class BaseHTTPAdapter:
+class BaseHTTPAdapter(ABC):
     client = None
 
+    @abstractmethod
     def get_with_data(self, url, data):
         raise NotImplementedError
 
+    @abstractmethod
     def post(self, url, data):
         raise NotImplementedError
 
 
-class AsyncHTTPAdapter(BaseHTTPAdapter):
-    pass
-
-
 class SyncHTTPAdapter(BaseHTTPAdapter):
     def get_with_data(self, url, data):
-        logger.debug(f"[SyncHTTPAdapter:get_with_data] URL: {url}")
+        extra = {"url": url}
+        logger.debug("[django_athm:get_with_data]", extra=extra)
 
         with httpx.Client(base_url=API_BASE_URL) as client:
             response = client.request(method="GET", url=url, json=data)
             return response.json()
 
     def post(self, url, data):
-        logger.debug(f"[SyncHTTPAdapter:post] URL: {url}")
+        extra = {"url": url}
+        logger.debug("[django_athm:post]", extra=extra)
 
         with httpx.Client(base_url=API_BASE_URL) as client:
             response = client.post(url, json=data)
@@ -42,5 +43,4 @@ class SyncHTTPAdapter(BaseHTTPAdapter):
 
 
 def get_http_adapter():
-    # TODO: If async is supported, use the AsyncHTTPAdapter
     return SyncHTTPAdapter()