+++ tests/conftest.py
@@ -2,7 +2,7 @@ import pytest
 import respx
 from httpx import Response
 
-from django_athm.constants import API_BASE_URL, LIST_URL, REFUND_URL, STATUS_URL
+from django_athm.constants import API_BASE_URL, REFUND_URL, REPORT_URL, SEARCH_URL
 from django_athm.utils import SyncHTTPAdapter
 
 
@@ -19,10 +19,10 @@ def mock_http_adapter_post(mocker):
 @pytest.fixture
 def mock_httpx():
     with respx.mock(base_url=API_BASE_URL, assert_all_called=False) as respx_mock:
-        get_with_data_route = respx_mock.get(LIST_URL, name="get_with_data")
+        get_with_data_route = respx_mock.get(REPORT_URL, name="get_with_data")
         get_with_data_route.return_value = Response(200, json={"mocked": True})
 
-        post_status_route = respx_mock.post(STATUS_URL, name="post_status", json={})
+        post_status_route = respx_mock.post(SEARCH_URL, name="post_status", json={})
         post_status_route.return_value = Response(200, json={"mocked": True})
 
         post_refund_route = respx_mock.post(REFUND_URL, name="post_refund", json={})